Refactor imports, test setup and remove snapshot dependencies
