+++ pom.xml
@@ -30,13 +30,13 @@
     <dependency>
       <groupId>uk.gov.ons.ctp.product</groupId>
       <artifactId>partysvc-api</artifactId>
-      <version>10.50.6-SNAPSHOT</version>
+      <version>10.50.6</version>
     </dependency>
 
     <dependency>
       <groupId>uk.gov.ons.ctp.product</groupId>
       <artifactId>samplesvc-api</artifactId>
-      <version>10.49.10-SNAPSHOT</version>
+      <version>10.49.10</version>
     </dependency>
 
     <dependency>
